Rename cx to classNames
